time_series renamed as tsa
its poor convention to have underscores in module names, and the
identical naming of the module and the time_series class was causing
some confusion. It is best to just change the name of one of them. so
the time_series module has been renamed as "tsa" for "time series
analysis"